feat: re-add smart motor controller group
This partially reverts commit 24c81a39cb90bf4b1ac4c1abf660085011cddd38.
It extends the Smart Motor Controller Group to pass all functionality to the first motor controller. This requires that the additional motors are configured to follow the primary motor outside of this class.

This class is primarily meant to facilitate logging of current and temperature data for multiple motors in a follower group.